Argument dummy

Argument dummy

argument_dummy(df, normalize, verbose)

Arguments

df

A data.frame.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

across

Character value. Specifies the discrete variable in the data.frame across which the variables of interest are to be analyzed or displayed.

across.subset

Character vector. The groups of interest that the grouping variable denoted in across contains.

If there are more groups you are interested in than they are groups you are not interested in specify those that you are not interested in prefixed with an '-'. Variables prefixed that way are discarded and the remaining are kept.

alpha.by, alpha_by, shape.by, shape_by, size.by, size_by

Character value or NULL. If character, specifies the variable that is mapped to the respective aesthetic of the plot.

alpha.trans, alpha_trans, color.trans, color_trans, shape.trans, shape_trans, size.trans, size_trans

Character value. Name of the transformation method applied to the continuous scale. Use validScaleTransformations() to obtain all valid input options. Use 'identity' for no transformation.

color.by, color_by

Character value or NULL. If character, specifies the variable that is displayed by color.

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrsp, pt.clrsp, pt_clrsp

Character value. Specifies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

clrp.adjust, clrp_adjust

Named character vector. Can be used to adjust colors with which groups are displayed. Names indicate groups, values indicate the colors with which specific groups are supposed to be displayed.

color.aes, color_aes

Either 'color' or 'fill'. Specifies the aesthetic that is used to visualize the variable that is specified in color_by. Ignored if color_by is NULL.

display_labels

Lgoical value. Indicates if labels are displayed.

display_legend, display_title

Logical values. Indicate if legend or title is displayed.

display.points

Logical value. If set to TRUE points are used additionally to display the results.

display.facets

Logical value. If set to TRUE the plot is split via ggplot2::facet_wrap() such that each variable gets it's own subplot.

display.grid, display_grid

Logical value. If TRUE, a grid is displayed.

facet.with, facet_with

Character value. Either 'grid' or 'wrap'. Specifies the function with which the plot-facetting is created. If across is of length 2 and facet_with = 'wrap' ggplot2::facet_wrap() is used. Else ggplot2::facet_grid() is used.

force

Logical value. Must be set to TRUE to allow overwriting.

grid.alpha, grid_alpha, grid.size, grid_size

Numeric values. Specify transparency and thickness of the lines of the grid.

grid.color, grid_color

Character value. Specifies the color of the grid lines.

h

Numeric value. Denotes the height at which the dendrogram is cut.

hs

Numeric vector. Denotes the heights at which the dendrogram is cut.

key.name, key_name

Character value or NULL. Denotes the variable that is used to identify each observation uniquely.

key.prefix, key_prefix

Character value. The string with which the IDs in the newly constructed key variable are prefixed if key_name is NULL.

k

Numeric value. Denotes the number of clusters. Must be bigger than 1. Must not be bigger than the number of observations of the data set.

ks

Numeric vector. Denotes all options for k-clusters. Values <1 are discarded. Is converted into an integer vector.

labels.angle, labels_angle

Numeric value. The angle with which labels are displayed.

labels.size, labels_size

Numeric value. The size with which labels are displayed.

labels.hjust, labels.vjust, labels_hjust, labels_vjust

Numeric values. Adjust labels positioning.

line.alpha, line_alpha, line.size, line_size

Numeric value. Specifies the transparency and the thickness of lines in the plot.

line.color, line_color

Character value. Specifies the color of lines in the plot.

line.type, line_type

Character value. Specifies the type of the lines. Use validLineTypes() to obtain all valid input options.

method.aggl, method_aggl

Character value. Specifies the agglomerative method of interest. Use validMethodsAggl() to obtain all valid input options.

methods.aggl, methods_aggl

Character vector. Specifies the agglomerative methods of interest. Use validMethodsAggl() to obtain all valid input options.

method.corr, method_corr

Character value. Specifies the correlation method of interest. Use validMethodsCorrelation() to obtain all valid input options.

methods.corr, methods_corr

Character vector. Specifies the correlation methods of interest. Use validMethodsCorrelation() to obtain all valid input options.

method.dist, method_dist

Character value. Specifies the distance method of interest. Use validMethodsDist() to obtain all valid input options.

methods.dist, methods_dist

Character vector. Specifies the distance methods of interest. Use validMethodsDist() to obtain all valid input options.

method.kmeans, method_kmeans

Character value. Specifies the kmeans method of interest. Use validMethodsKmeans() to obtain all valid input options.

methods.kmeans, methods_kmeans

Character vector. Specifies the kmeans methods of interest. Use validMethodsKmeans() to obtain all valid input options.

method.pam, method_pam

Character value. Specifies the pam-method of interest. Use validMethodsPam() to obtain all valid input options.

methods.pam, methods_pam

Character value. Specifies the pam-method of interest. Use validMethodsPam() to obtain all valid input options.

object

Any object for whose class a method has been defined.

order.by, order_by

Character value or NULL. If character, specifies the numeric variable by which the points are arranged before plotting.

order.desc

Logical value. If set to TRUE the way the points are ordered is reversed.

pt.alpha, pt_alpha

Numeric value. Specifies the transparency of points.

pt.color, pt_color

Character value. Specifies the color with which all points are displayed .

pt.num, pt_num

Numeric value. Species the number of points (sample size) to prevent overplotting.

pt.shape, pt_shape

Numeric or character value. If numeric, the respective shape is taken for all points. If character, the respective variable is mapped to the shape-aesthetic.

pt.size, pt_shape

Numeric values. Specifies size of points.

pval_threshold

Numeric value or NULL. If numeric, denotes the maximum p-value to be considered as significant. Must no be higher than 0.05.

ref.group

Character value. Specifies the reference group for the pairwise statistical test. Must be among the groups the variable specified in across contains. If set to NULL the first group found is taken.

relevel

Logical value. If set to TRUE the input order of across.subset and variables determines the order in which the groups of interest are displayed.

scales, nrow, ncol

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

smooth.alpha, smooth_alpha

Numeric value. The transparency of the smoothed line.

smooth.color, smooth_color

Character value. The color of the smoothed line.

smooth.method, smooth_method

Character value. The smoothing method. Given to argument method of ggplot2::geom_smooth().

smooth.se, smooth_se

Logical value. Indicates if the confidence intervals are displayed.

stop_if_null

Logical value. If TRUE and the function does not find the object to extract an informative error is raised. Else the empty value - usually NULL - is returned.

transform.with, transform_with

List or NULL. If list, can be used to transform continuous variables before plotting. Names of the slots of the provided list refer to the variables. The content of the slot refers to the transforming functions. Slot content can either be a character vector of function names. Use validVarTransformations() to obtain all valid character value inputs. Or it can be a list of functions (and function names).

test.pairwise

Character value or NULL. If character, one of 't.test', 'wilcox.test'.

test.groupwise

Character value or NULL. If character, one of 'anova', 'kruskal.test'.

valid.classes

Character vector to specify the classes the input might have in order not to be discarded.

values_alpha, values_size

Numeric values. Denote transparency and size of the displayed values.

values_color

Character value. Denotes the color in which the values are displayed.

values_digits

Numeric value. Denotes the number of digits to which the displayed values are rounded.

variables

Character vector. Specifies the variables of interest. If set to NULL all valid variables of the input data.frame are considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables you are not interested in specify those that you are not interested in prefixed with an '-'. Variables prefixed that way are discarded and the remaining are kept.

x, y

Character value. Variable to be plotted on the respective axis.

Value

Returns a ggplot-object that can be additionally customized according to the rules of the ggplot2-framework.