corr_dummy

corr_dummy

corr_dummy(diagonal, digits, distinct, type)

Arguments

diagonal

Logical value. Indicates if the diagonal values of the matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if type = 'complete'.

digits

Numeric. Given to base::round() and indicates the number of digits to which the correlation value is rounded. Defaults to 2.

distinct

Logical value. If TRUE only one observation per variable pair remains in the output data.frame. Ignored if type = 'complete'.

type

Character value. Denotes how the underlying correlation matrix is handled. Three options:

'complete': The matrix stays as is.,

'lower': The part below the diagonal is used. The upper part is set to NA.,

'upper':The part above the diagonal is used. the lower part is set to NA.

shape

Character value. Specifies the shape of the geometric objects with which the variable pairs are displayed. Either 'circle', 'rect' or 'tile'.