Search for genes

Goes through the summary text of genes and returns gene names that match the requirements.

filter_genes(
  catchphrases,
  test = "any",
  standalone = TRUE,
  case_sensitive = FALSE,
  return_var = "symbol",
  genes_subset = NULL,
  genes_rm = NULL,
  ...
)

Arguments

catchphrases

Character vector of catchphrases to look for.

test

Character value. Determines how the input of catchphrases is used to skim the genes.

test = 'any': Genes are kept if at least one of the catchphrases appears in the summary.

test = 'all': Genes are kept if all of the catchphrases appear in the summary.

test = 'none': Genes are kept if none of the catchphrases appear in the summary

standalone

Logical value. If TRUE, catchphrases must appear as standalone words. If FALSE, they can be part of other words.

case_sensitive

Logical value. If TRUE, the function pays attention to capital- and non-capital letters. E.g. if catchphrases = 'astrocyte' but the text of a gene only features Astrocyte the gene is not included in the output as the text features astrocyte in upper case. Setting case_sensitive to FALSE, the default, makes the function ignore upper and lower cases.

return_var

The variable of the gene_info_df that is returned as a vector. Defaults to symbol.

genes_subset

Character vector or NULL. If character, subsets gene info data.frame accordingly such that the test is only applied to the provided genes.

genes_rm

Character vector or NULL. If character, removes specified genes from gene info data.frame such that the catchphrase test is only applied to genes that are not specified.

...

Logical tests given to dplyr::filter() that can be used to prefilter the data.frame before going through the summary texts.

Value

Character vector of gene names.

Details

See documentation for ?gene_info_df to read about how the source data.frame has been generated.

Examples

 # create a small subset of genes for demonstration purposes

 example_genes <- c("GFAP" , "MAG", "OLIG1", "EGFR")

 print_gene_info(genes = example_genes)
#> --------------------------------------------------
#> Name: glial fibrillary acidic protein (GFAP)
#> Type: protein coding
#> Synonyms: ALXDRD
#> 
#> This gene encodes one of the major intermediate filament proteins of mature astrocytes. It is used as a marker to distinguish astrocytes from other glial cells during development. Mutations in this gene cause Alexander disease, a rare disorder of astrocytes in the central nervous system. Alternative splicing results in multiple transcript variants encoding distinct isoforms.
#> --------------------------------------------------
#> Name: myelin associated glycoprotein (MAG)
#> Type: protein coding
#> Synonyms: GMA; S-MAG; SPG75; SIGLEC4A; SIGLEC-4A
#> 
#> The protein encoded by this gene is a type I membrane protein and member of the immunoglobulin superfamily. It is thought to be involved in the process of myelination. It is a lectin that binds to sialylated glycoconjugates and mediates certain myelin-neuron cell-cell interactions. Three alternatively spliced transcripts encoding different isoforms have been described for this gene.
#> --------------------------------------------------
#> Name: oligodendrocyte transcription factor 1 (OLIG1)
#> Type: protein coding
#> Synonyms: BHLHB6; BHLHE21
#> 
#> Predicted to enable DNA-binding transcription factor activity, RNA polymerase II-specific and RNA polymerase II cis-regulatory region sequence-specific DNA binding activity. Predicted to be involved in neuron differentiation and regulation of transcription by RNA polymerase II. Predicted to act upstream of or within neuron fate commitment. Predicted to be part of chromatin. Predicted to be active in nucleus.
#> --------------------------------------------------
#> Name: epidermal growth factor receptor (EGFR)
#> Type: protein coding
#> Synonyms: ERBB; ERRP; HER1; mENA; ERBB1; PIG61; NISBD2
#> 
#> The protein encoded by this gene is a transmembrane glycoprotein that is a member of the protein kinase superfamily. This protein is a receptor for members of the epidermal growth factor family. EGFR is a cell surface protein that binds to epidermal growth factor, thus inducing receptor dimerization and tyrosine autophosphorylation leading to cell proliferation. Mutations in this gene are associated with lung cancer. EGFR is a component of the cytokine storm which contributes to a severe form of Coronavirus Disease 2019 (COVID-19) resulting from infection with severe acute respiratory syndrome coronavirus-2 (SARS-CoV-2).

 # different tests

 filter_genes(
  catchphrases = c("astrocytes", "surface"),
  test = "any",
  genes_subset = example_genes
  )
#> [1] "EGFR" "GFAP"

 filter_genes(
  catchphrases = c("astrocytes", "surface"),
  test = "all",
  genes_subset = example_genes
  )
#> character(0)

 filter_genes(
  catchphrases = c("astrocytes", "surface"),
  test = "none",
  genes_subset = example_genes
  )
#> [1] "MAG"   "OLIG1"

 # standalone TRUE or FALSE

 filter_genes(
  catchphrases = c("glyco"),
  standalone = FALSE,
  genes_subset = example_genes
  )
#> [1] "EGFR" "MAG" 

 filter_genes(
  catchphrases = c("glyco"),
  standalone = TRUE,
  genes_subset = example_genes
  )
#> character(0)