Collection of gene information

Collection of gene information stored in tidy data fashion. Information has been retrieved from the gene data base of the National Library of Medicine. Lastly updated on 27.10.2022.

data(gene_info_df)

Format

An object of class tbl_df (inherits from tbl, data.frame) with 25976 rows and 7 columns.

Source

https://www.ncbi.nlm.nih.gov/gene/?term=

id

Integer. Gene ID in the data base. E.g. gene GFAP has ID 2670. Combining them with the url https://www.ncbi.nlm.nih.gov/gene/id results in the url of the webpage from which the from which the information has been retrieved. E.g. https://www.ncbi.nlm.nih.gov/gene/2760

symbol

Character. HGNC symbol of the gene. (Common gene abbreviation.)

full_name

Character. Full gene name.

synonyms

Character. Synonyms with which to refer to the gene. Separated by ';'

summary

Character. Human readable summary text of the gene. If NA, no summary has been found on the webpage.

type

Character. Describes the general function of the gene. E.g. protein coding, pseudo, scRNA, snoRNA, snRNA, tRNA, etc.

organism

Character. The organism in which the gene appears. Currently only Homo sapiens.