Obtain object of class rcorr

Extracts correlation results in form of class rcorr.

getRcorr(object, ...)

# S4 method for Correlation
getRcorr(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE,
  flatten = TRUE
)

# S4 method for CorrelationMethod
getRcorr(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  as_list = FALSE,
  stop_if_null = TRUE
)

Arguments

object

Any object for whose class a method has been defined.

method_corr

Character value. Specifies the correlation method of interest. Use validMethodsCorrelation() to obtain all valid input options.

across

Character value. Specifies the discrete variable in the data.frame across which the variables of interest are to be analyzed or displayed.

as_list

Logical value. If TRUE, rcorr objects are returned as simple lists.

stop_if_null

Logical value. If TRUE and the function does not find the object to extract an informative error is raised. Else the empty value - usually NULL - is returned.

Value

rcorr returns a list with elements r, the matrix of correlations, n the matrix of number of observations used in analyzing each pair of variables, and P, the asymptotic P-values. Pairs with fewer than 2 non-missing values have the r values set to NA. The diagonals of n are the number of non-NAs for the single variable corresponding to that row and column.

Details

If across is NULL one object of class rcorr is returned - the one for the complete data set. If across is a character the list of rcorr is extracted and subsetted by the groups of interest denoted in across_subset. If across_subset is NULL the complete list is returned. If across_subset is of length one and flatten is TRUE the output list is flattened and a single rcorr object is returned.