Glue a human readable list report

Combines all slots of the specified list in lst that are values to a character/glue object.

glue_list_report(lst, prefix = "", separator = " = ", combine_via = "\n", ...)

Arguments

lst

A named list of values.

prefix

Character value. Denotes the string with which to prefix each slots.

separator

Character value or NULL. Denotes the string with which the slot name is combined with the slot's content. If set to NULL neither the slot names nor the separators are mentioned and the slot's contents are combined as they are.

Value

Glue object.

Examples

#Not run:

 lst_input <- list("arg1" = TRUE, "arg2" = glue::glue_collapse(1:5, sep = ", "))

 glue_list_report(lst = lst_input, separator = " = ")
#> 
#> arg1 = TRUE
#> arg2 = 1, 2, 3, 4, 5