Plot boxplot

Plots a boxplot

plotBoxplot(object, ...)

# S4 method for Analysis
plotBoxplot(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  vjust = 0,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  pt_alpha = 0.8,
  pt_clr = "black",
  pt_num = 100,
  pt_size = 1.25,
  pt_shape = 21,
  verbose = TRUE,
  ...
)

# S4 method for Analysis
plotViolinplot(
  object,
  variables,
  phase = NULL,
  across = NULL,
  across_subset = NULL,
  relevel = TRUE,
  clrp = "milo",
  clrp_adjust = NULL,
  test_groupwise = NULL,
  test_pairwise = NULL,
  ref_group = NULL,
  step_increase = 0.01,
  vjust = 0,
  display_facets = TRUE,
  scales = "free",
  nrow = NULL,
  ncol = NULL,
  display_points = FALSE,
  pt_alpha = 0.8,
  pt_clr = "black",
  pt_num = 100,
  pt_size = 1.25,
  pt_shape = 21,
  verbose = TRUE,
  ...
)

Arguments

object

Any object for whose class a method has been defined.

variables

Character vector. Specifies the variables of interest. If set to NULL all valid variables of the input data.frame are considered (as long as the total number does not exceeds the limit).

If there are more variables you are interested in than they are variables you are not interested in specify those that you are not interested in prefixed with an '-'. Variables prefixed that way are discarded and the remaining are kept.

across

Character value. Specifies the discrete variable in the data.frame across which the variables of interest are to be analyzed or displayed.

relevel

Logical value. If set to TRUE the input order of across.subset and variables determines the order in which the groups of interest are displayed.

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

scales, nrow, ncol

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

A ggplot.