Plot correlation plot

Plot correlation plot

plotCorrplot(object, ...)

# S4 method for Correlation
plotCorrplot(
  object,
  method_corr = "pearson",
  across = NULL,
  across_subset = NULL,
  variables_subset = NULL,
  relevel = FALSE,
  pval_threshold = NULL,
  type = "lower",
  diagonal = TRUE,
  color_low = "darkred",
  color_high = "steelblue",
  shape = "tile",
  size_by_corr = TRUE,
  size_max = 15,
  display_values = TRUE,
  values_alpha = 0.9,
  values_color = "black",
  values_digits = 2,
  values_size = 4,
  display_grid = TRUE,
  grid_color = "grey",
  grid_size = 0.5,
  nrow = NULL,
  ncol = NULL,
  verbose = TRUE
)

Arguments

object

Any object for whose class a method has been defined.

method_corr

Character value. Specifies the correlation method of interest. Use validMethodsCorrelation() to obtain all valid input options.

across

Character value or NULL. If character, specifies the name of the grouping variable that contains the group assignment across which results are displayed. If NULL, no grouping is done.

across_subset

Character vector or NULL. If character, specifies groups of the grouping variable that are of interest. Prefixing group names with '-' excludes them. If NULL, all groups are considered.

variables_subset

Character vector or NULL. If character, specifies the subset of variable names that is included in the plot. To exclude single variables the variable name can be provided prefixed with a '-'.

relevel

Logical value. If set to TRUE the input order of across.subset and variables determines the order in which the groups of interest are displayed.

pval_threshold

Numeric value or NULL. If numeric, denotes the maximum p-value to be considered as significant. Must no be higher than 0.05.

type

Character value. Denotes how the underlying correlation matrix is handled. If type = 'complete', the matrix is let as is. If type = 'lower', the part below the diagonal is used. The upper part is set to NA. If type = 'upper, the part above the diagonal is used. the lower part is set to NA.

diagonal

Logical value. Indicates if the diagonal values of the matrices are kept ( = TRUE) or set to NA (= FALSE). Ignored if type = 'complete'.

color_low, color_high

Character values. Specifies the colors that are used to generate the color spectrum with which the gradient from negative color_low) to positive correlation (color_high) is displayed.

shape

Character value. Specifies the shape of the geometric objects with which the variable pairs are displayed. Either 'circle', 'rect' or 'tile'.

size_by_corr

Logical value. If TRUE, size is used in addition to coloring to indicate the correlation. Ignored if shape = 'tile'

values_alpha

Numeric values. Denote transparency and size of the displayed values.

values_color

Character value. Denotes the color in which the values are displayed.

values_digits

Numeric value. Denotes the number of digits to which the displayed values are rounded.

values_size

Numeric values. Denote transparency and size of the displayed values.

display_grid

Logical value. If TRUE, a grid is displayed.

grid_color

Character value. Specifies the color of the grid lines.

grid_size

Numeric values. Specify transparency and thickness of the lines of the grid.

nrow

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

ncol

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

verbose

Logical. If set to TRUE informative messages regarding the computational progress will be printed.

(Warning messages will always be printed.)

Value

A ggplot.