Plot dendrograms

Plots a dendrogram with either ggplot or base plot.

plotDendrogram(object, ...)

# S4 method for Clustering
plotDendrogram(
  object,
  methods_dist = "euclidean",
  methods_aggl = "ward.D",
  k = NULL,
  h = NULL,
  type = "rectangle",
  facet_with = "grid",
  direction = "bt",
  branch_color = "black",
  branch_size = 1,
  display_labels = FALSE,
  labels_angle = 90,
  labels_hjust = 0,
  labels_nudge = 0.01,
  labels_size = 3,
  labels_vjust = 0.5,
  display_legend = TRUE,
  display_title = FALSE,
  clrp = "milo",
  clrp_adjust = NULL,
  simple = FALSE,
  nrow = NULL,
  ncol = NULL,
  ...
)

Arguments

object

Any object for whose class a method has been defined.

methods_dist

Character vector. Specifies the distance methods of interest. Use validMethodsDist() to obtain all valid input options.

methods_aggl

Character vector. Specifies the agglomerative methods of interest. Use validMethodsAggl() to obtain all valid input options.

k

Numeric value. Denotes the number of clusters. Must be bigger than 1. Must not be bigger than the number of observations of the data set.

h

Numeric value. Denotes the height at which the dendrogram is cut.

type

Character value. Indicates the shape of the dendrogram. Input 'rectangle' will draw rectangular lines, while 'triangle' will draw triangular lines.

facet_with

Character value. Either 'grid' or 'wrap'. Specifies the function with which the plot-facetting is created. If the number of input combinations for methods_dist and methods_aggl length 2 or bigger and facet_with = 'wrap' ggplot2::facet_wrap() is used. Else ggplot2::facet_grid() is used.

display_labels

Lgoical value. Indicates if labels are displayed.

labels_angle

Numeric value. The angle with which labels are displayed.

labels_hjust

Numeric values. Adjust labels positioning.

labels_size

Numeric value. The size with which labels are displayed.

labels_vjust

Numeric values. Adjust labels positioning.

display_legend

Logical values. Indicate if legend or title is displayed.

display_title

Logical values. Indicate if legend or title is displayed.

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust

Named character vector. Can be used to adjust colors with which groups are displayed. Names indicate groups, values indicate the colors with which specific groups are supposed to be displayed.

simple

Logical value. If TRUE, the dendrogram is plotted with base::plot(). This is way quicker but does not allow for ggplot2 specific adjustments.

nrow

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

ncol

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

Value

A ggplot or a base plot.