Plot PCA

Plots PCA results in a scatterplot.

plotPCA(object, ...)

# S4 method for DimRed
plotPCA(
  object,
  n_dims = 2,
  alpha_by = NULL,
  color_by = NULL,
  shape_by = NULL,
  size_by = NULL,
  pt_alpha = 0.9,
  pt_color = "black",
  pt_fill = "black",
  pt_shape = 19,
  pt_size = 1,
  color_aes = "color",
  clrp = "milo",
  clrp_adjust = NULL,
  clrsp = "inferno",
  ...
)

Arguments

object

Any object for whose class a method has been defined.

n_dims

Numeric value. Integer that indicates the number of principal component vectors that are plotted.

alpha_by

Character value or NULL. If character, specifies the variable that is mapped to the respective aesthetic of the plot.

color_by

Character value or NULL. If character, specifies the variable that is displayed by color.

shape_by

Character value or NULL. If character, specifies the variable that is mapped to the respective aesthetic of the plot.

size_by

Character value or NULL. If character, specifies the variable that is mapped to the respective aesthetic of the plot.

pt_alpha

Numeric value. Specifies the transparency of points.

pt_color

Character value. Specifies the color with which all points are displayed .

pt_shape

Numeric or character value. If numeric, the respective shape is taken for all points. If character, the respective variable is mapped to the shape-aesthetic.

color_aes

Either 'color' or 'fill'. Specifies the aesthetic that is used to visualize the variable that is specified in color_by. Ignored if color_by is NULL.

clrp

Character value. Specifies the color palette to be used to represent groups of discrete variables. Run validColorPalettes() to obtain valid input options.

clrp_adjust

Named character vector. Can be used to adjust colors with which groups are displayed. Names indicate groups, values indicate the colors with which specific groups are supposed to be displayed.

clrsp

Character value. Specfies the color spectrum to be used to represent continuous values of numeric variables. Run validColorSpectra() to obtain valid input options.

Value

A ggplot.