Plot 1d dot plot

Plots a dot plot of one categorical variable. Numeric values can be displayed via alpha, color, size and x-axis.

plot_dot_plot_1d(
  df,
  x,
  y,
  across = NULL,
  across.subset = NULL,
  relevel = TRUE,
  reorder = TRUE,
  reorder.rev = FALSE,
  alpha.by = NULL,
  alpha.trans = "identity",
  color.by = NULL,
  color.trans = "identity",
  shape.by = NULL,
  size.by = NULL,
  size.trans = "reverse",
  pt.alpha = 0.9,
  pt.color = "black",
  pt.clrp = "default",
  pt.clrsp = "plasma",
  pt.shape = 19,
  pt.size = 3,
  scales = "free_y",
  nrow = NULL,
  ncol = NULL,
  transform.with = NULL,
  ...
)

Arguments

df

A data.frame.

x

Character value. The numeric continuous varable.

y

Charcter value. The grouping variable.

across

Character value. Specifies the discrete variable in the data.frame across which the variables of interest are to be analyzed or displayed.

across.subset

Character vector. The groups of interest that the grouping variable denoted in across contains.

If there are more groups you are interested in than they are groups you are not interested in specify those that you are not interested in prefixed with an '-'. Variables prefixed that way are discarded and the remaining are kept.

relevel

Logical value. If set to TRUE the input order of across.subset and variables determines the order in which the groups of interest are displayed.

scales, nrow, ncol

Given to ggplot2::facet_wrap(). Affects the way the subplots are displayed.

Details

A ggplot.